requirements.txt
discord.py<3.0.0
requests
aiohttp
pygame
python-datamuse
python-chess
googletrans
